# swagger.yaml
openapi: 3.0.3
info:
  title: Dating App API
  version: 1.0.0
  description: |
    Music-based dating application backend. Matches users based on Spotify listening history,
    lifestyle compatibility, and behavioral learning.

    ## Authentication
    Most endpoints require a JWT passed as `Authorization: Bearer <token>`.
    JWT claims: `{ userId, tokenVersion, sub, iat, exp }`.

    After a password reset `tokenVersion` is incremented — all prior tokens are invalidated.
    Deleted-user JWTs return **401** (not 404).

    ## Email Verification
    Unverified-email users receive **403** `{"error":"Email verification required"}` on all
    endpoints except `/api/v1/auth/**`, `/api/v1/onboarding/**`, `/public/**`, and `/api/test/**`.

    ## Rate Limiting
    | Action | Limit | Key |
    |--------|-------|-----|
    | Swipe | 60/min | userId |
    | Score calculation | 30/min | userId |
    | Unauthenticated | per IP | remoteAddr |

    Exceeded limits return **429**.

    ## Error Contract
    Every error response follows:
    ```json
    { "code": "string", "message": "string", "fields": {}, "timestamp": "ISO8601" }
    ```
    Always switch on `code`, never on `message`.

  contact:
    name: API Support
    email: mladen.hangi@kodelab.hr
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/matching/score/{otherUserId}:
    get:
      tags: [Matching]
      summary: Get detailed compatibility score with another user
      description: |
        Returns the directional score (current user → other user). Rate limited to 30/min.
        Blocked relationships return 403. Self-scoring returns 400.
      operationId: getMatchScore
      parameters:
        - name: otherUserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440001
      responses:
        '200':
          description: Detailed compatibility score
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchScore'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/matching/swipe:
    post:
      tags: [Matching]
      summary: Record a swipe action
      description: |
        Rate limited to 60/min.
        - `super_like` triggers instant match.
        - `block` prevents all future interactions and auto-unmatches any active match.
      operationId: recordSwipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwipeRequest'
            example:
              swipedUserId: 550e8400-e29b-41d4-a716-446655440001
              action: like
              platform: web
      responses:
        '200':
          description: Swipe recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwipeResult'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/DuplicateSwipe'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /api/v1/preferences/genres:
    get:
      tags: [Genre Preferences]
      summary: Get current user's genre preferences
      operationId: getGenrePreferences
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Genre preferences list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenrePreferenceListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Genre Preferences]
      summary: Add a manual genre preference
      operationId: addGenrePreference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGenrePreferenceRequest'
            example:
              genreName: indie-rock
              weight: 0.9
      responses:
        '201':
          description: Genre preference added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenrePreferenceCreateResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/preferences/genres/{genreName}:
    delete:
      tags: [Genre Preferences]
      summary: Remove a specific genre preference
      operationId: removeGenrePreference
      parameters:
        - name: genreName
          in: path
          required: true
          schema:
            type: string
          example: indie-rock
      responses:
        '200':
          description: Genre preference removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageResponse'
              example:
                success: true
                message: Genre preference removed
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ─── Account ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT with claims: `{ userId, tokenVersion, sub, iat, exp }`.
        `tokenVersion` is required — tokens without it are rejected with 401.

  # ─── Responses ─────────────────────────────────────────────────────────────
  responses:
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: Validation failed
            fields:
              email: must not be blank
              password: Password must contain at least one uppercase letter, one lowercase letter, one number, and one special character
            timestamp: '2026-05-12T10:15:30.123Z'
    InvalidArgument:
      description: Invalid query parameter or illegal request state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INVALID_ARGUMENT
            message: Invalid status parameter. Allowed values active, all
            timestamp: '2026-05-12T10:15:30.123Z'
...
    Unauthorized:
      description: Missing or invalid JWT, or wrong credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Invalid email or password
            timestamp: '2026-05-12T10:15:30.123Z'
    Forbidden:
      description: Authenticated but not authorized for this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: Access denied
            timestamp: '2026-05-12T10:15:30.123Z'
...
    DuplicateSwipe:
      description: User has already swiped on this person
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: DUPLICATE_SWIPE
            message: You have already swiped on this user
            timestamp: '2026-05-12T10:15:30.123Z'
...
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: User not found
            timestamp: '2026-05-12T10:15:30.123Z'
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: ACCOUNT_LOCKED
            message: Too many requests. Please slow down.
            timestamp: '2026-05-12T10:15:30.123Z'
...
  schemas:
...
    GenrePreferenceSource:
      type: string
      description: Serialized as lowercase in JSON
      enum: [spotify_derived, manual_selection, inferred, hybrid, seed_data]
    CompatibilityLevel:
      type: string
      enum: [LOW, MEDIUM, HIGH, VERY_HIGH]
      description: LOW=0-40, MEDIUM=41-60, HIGH=61-80, VERY_HIGH=81-100

    # ── Error ──────────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
          description: Machine-readable error code. Always switch on this, never on message.
          example: VALIDATION_ERROR
          enum:
            - VALIDATION_ERROR
            - INVALID_ARGUMENT
            - INVALID_TOKEN
            - UNAUTHORIZED
            - FORBIDDEN
            - ACCOUNT_LOCKED
            - NOT_FOUND
            - EMAIL_EXISTS
            - CONFLICT
            - DUPLICATE_SWIPE
            - CONCURRENT_MODIFICATION
            - SPOTIFY_TOKEN_EXPIRED
            - SPOTIFY_UNAVAILABLE
            - INTERNAL_ERROR
        message:
          type: string
          description: Human-readable description
        fields:
          type: object
          nullable: true
          description: Per-field validation messages. Only present for VALIDATION_ERROR.
          additionalProperties:
            type: string
          example:
            email: must not be blank
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp

    # ── Auth ───────────────────────────────────────────────────────────────
...
    SharedGenre:
      type: object
      properties:
        genre:
          type: string
          example: indie-rock
        genreDisplayName:
          type: string
          example: Indie Rock
        userWeight:
          type: number
          format: double
          description: Requesting user's weight for this genre (0–1)
          example: 0.85
        otherWeight:
          type: number
          format: double
          description: Other user's weight for this genre (0–1)
          example: 0.72
        overlap:
          type: number
          format: double
          description: min(userWeight, otherWeight)
          example: 0.72
        similarity:
          type: number
          format: double
          description: 1.0 = identical weights, 0.0 = very different
          example: 0.85
    MatchBreakdown:
      type: object
      properties:
        sharedGenres:
          type: array
          items:
            $ref: '#/components/schemas/SharedGenre'
        userOnlyGenres:
          type: array
          items:
            type: string
        otherOnlyGenres:
          type: array
          items:
            type: string
        sharedGenreCount:
          type: integer
          example: 3
        totalUniqueGenres:
          type: integer
          example: 8
        genreOverlapScore:
          type: number
          format: double
          example: 75.0
        weightSimilarityScore:
          type: number
          format: double
          example: 68.0
        diversityScore:
          type: number
          format: double
          example: 55.0
        matchConfidence:
          type: number
          format: double
          example: 0.8
        sharedInterests:
          type: array
          items:
            type: string
          example: [hiking, concerts]
        lifestyleCompatibilityScore:
          type: number
          format: double
          example: 80.0
    MatchScore:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        otherUserId:
          type: string
          format: uuid
        overallScore:
          type: number
          format: double
          description: 0–100
          example: 78.4
        musicScore:
          type: number
          format: double
          description: 0–100
          example: 85.0
        lifestyleScore:
          type: number
          format: double
          description: 0–100
          example: 72.0
        interestsScore:
          type: number
          format: double
          description: 0–100
          example: 65.0
        locationScore:
          type: number
          format: double
          description: 0–100
          example: 90.0
        behavioralScore:
          type: number
          format: double
          nullable: true
          description: 0–100; null during cold start (fewer than 5 likes)
          example: 70.0
        breakdown:
          $ref: '#/components/schemas/MatchBreakdown'
        insights:
          type: array
          items:
            type: string
          example:
            - You share 3 genres including indie-rock and alternative
            - Both value concerts highly
        compatibilityLevel:
          $ref: '#/components/schemas/CompatibilityLevel'
        calculatedAt:
          type: string
          format: date-time
          example: '2026-05-12T10:15:30.000Z'
    SwipeRequest:
      type: object
      required: [swipedUserId, action]
      properties:
        swipedUserId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440001
        action:
          type: string
          enum: [like, pass, super_like, block]
          example: like
        matchScore:
          type: number
          format: double
          nullable: true
          deprecated: true
          description: Deprecated — server computes the score internally; field is silently ignored
        platform:
          type: string
          maxLength: 50
          nullable: true
          enum: [web, ios, android]
          default: web
    SwipeResult:
      type: object
      properties:
        swipeId:
          type: string
          format: uuid
        action:
          type: string
          enum: [like, pass, super_like, block]
          example: like
        matchScore:
          type: number
          format: double
          description: Server-computed score at swipe time
          example: 78.4
        resultedInMatch:
          type: boolean
          example: false
        match:
          $ref: '#/components/schemas/SwipeMatchDetails'
          nullable: true
          description: Only present when resultedInMatch=true
    SwipeMatchDetails:
      type: object
      properties:
        matchId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          description: The other user's ID
        name:
          type: string
          example: Bob Jones
        matchScore:
          type: number
          format: double
          example: 78.4
        matchedAt:
          type: string
          format: date-time
          example: '2026-05-12T10:16:00.000Z'
...
    GenrePreferenceItem:
      type: object
      properties:
        genreName:
          type: string
          example: indie-rock
        genreDisplayName:
          type: string
          example: Indie Rock
        weight:
          type: number
          format: double
          description: 0.0–1.0
          example: 0.85
        confidence:
          type: number
          format: double
          example: 0.9
        source:
          $ref: '#/components/schemas/GenrePreferenceSource'
        rank:
          type: integer
          example: 1
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    GenrePreferenceListResponse:
      type: object
      properties:
        total:
          type: integer
          example: 12
        preferences:
          type: array
          items:
            $ref: '#/components/schemas/GenrePreferenceItem'
    AddGenrePreferenceRequest:
      type: object
      required: [genreName]
      properties:
        genreName:
          type: string
          example: indie-rock
        weight:
          type: number
          format: double
          minimum: 0.0
          maximum: 1.0
          default: 1.0
          nullable: true
    GenrePreferenceCreateResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Genre preference added
        preference:
          $ref: '#/components/schemas/GenrePreferenceItem'
...
    SuccessMessageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string